Get Changes from orig Project older than 15-Oct-2025

+++ esphome/components/uponor_smatrix145/climate/uponor_smatrix_climate.cpp
@@ -1,7 +1,7 @@
 #include "uponor_smatrix_climate.h"
+#include "esphome/core/application.h"
 #include "esphome/core/helpers.h"
 #include "esphome/core/log.h"
-#include "esphome/core/application.h"
 
 namespace esphome {
 namespace uponor_smatrix145 {
@@ -97,7 +97,7 @@ void UponorSmatrixClimate::send_data(float temperature, bool setByController = f
 }
 
 void UponorSmatrixClimate::on_device_data(const UponorSmatrixData *data, size_t data_len) {
-  for (int i = 0; i < data_len; i++) {
+  for (size_t i = 0; i < data_len; i++) {
     switch (data[i].id) {
       case UPONOR_ID_TARGET_TEMP_MIN:
         this->min_temperature_ = raw_to_celsius(data[i].value);
